refactor dcgan
