--- deeplay/components/dcgan/__init__.py
@@ -1,1 +1,0 @@
-from .dcgan import DCGANGenerator, DCGANDiscriminator
